requirements.txt
streamlit==1.29.0
requests==2.31.0
numpy==1.26.2
matplotlib==3.8.2
pandas==2.1.4
streamlit-lottie==0.0.5
scikit-learn==1.3.2
seaborn==0.13.0
tensorflow==2.7.0
pillow==9.4.0
opencv-python==4.8.1.78